+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - New: Japanese translation (Thanks momo-i!)
 - New: ProjectE support (all reinforced blocks now have an EMC value)
+- New: Taser effects and damage are now configurable
 - Fix: Nether portals can replace reinforced blocks when generating
 - Fix: Can't place blocks on blocks locked by a Sonic Security System
 - Fix: Interacting with blocks while holding a Sonic Security System may sometimes place it for a short while